# apps/web/openapi.yaml
openapi: 3.1.0
info:
  title: CountMeIn API
  description: API for group booking, organizer cabinet management, and notifications.
  version: 1.0.0+5fab21c281ec
servers:
  - url: https://countmein.group
    description: Production
  - url: http://localhost:3000
    description: Local development
paths:
  /api/auth/telegram-guest:
    post:
      summary: Verify Telegram login widget data and mint a single-use guest ticket
      operationId: telegramGuest
      x-rateLimit:
        limit: 10
        windowSeconds: 60
        per: ip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramWidgetPayload"
      responses:
        "200":
          description: Guest ticket issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestTicketResponse"
        "400":
          description: Malformed widget payload or failed HMAC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: Telegram bot is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/auth/telegram-signup:
    post:
      summary: Verify Telegram login widget data and mint an organizer signup ticket
      operationId: telegramSignup
      x-rateLimit:
        limit: 5
        windowSeconds: 60
        per: ip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramWidgetPayload"
      responses:
        "200":
          description: Auth ticket issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTicketResponse"
        "400":
          description: Malformed widget payload or failed HMAC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: Telegram bot is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
  /api/services:
    get:
      summary: List the services of the organizer this request may view
      operationId: listServices
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicesEnvelope"
        "500":
          description: Internal error
    post:
      summary: Create a service
      operationId: createService
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceEnvelope"
        "400":
          description: Validation error, or photoUrl outside the organizer media prefix
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidBody"
        "403":
          description: Demo account is read-only, or the caller has no session (ADR-010)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: Internal error
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Secure-authjs.session-token
      description: "Auth.js session cookie: `__Secure-authjs.session-token` in
        production, `authjs.session-token` in local development."
  schemas:
...
    AuthTicket:
      type: string
      minLength: 20
      maxLength: 200
...
    AuthTicketResponse:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/AuthTicket"
        organizerExists:
          type: boolean
      required:
        - ticket
        - organizerExists
...
    Capacity:
      type: integer
      minimum: 1
      maximum: 100000
    Contact:
      type: string
      minLength: 1
      maxLength: 300
...
    CreateServiceInput:
      type: object
      properties:
        title:
          $ref: "#/components/schemas/DisplayName"
        description:
          $ref: "#/components/schemas/ServiceDescription"
        location:
          $ref: "#/components/schemas/Location"
        contact:
          $ref: "#/components/schemas/Contact"
        defaultPrice:
          $ref: "#/components/schemas/PriceText"
        defaultCapacity:
          $ref: "#/components/schemas/Capacity"
        defaultDurationMinutes:
          $ref: "#/components/schemas/DurationMinutes"
        maxSeatsPerBooking:
          $ref: "#/components/schemas/MaxSeatsPerBooking"
        options:
          $ref: "#/components/schemas/OptionsList"
        optionsSelectMode:
          $ref: "#/components/schemas/OptionsSelectMode"
        photoUrl:
          $ref: "#/components/schemas/URL"
      required:
        - title
        - defaultPrice
        - defaultCapacity
        - defaultDurationMinutes
        - maxSeatsPerBooking
...
    DisplayName:
      type: string
      minLength: 1
      maxLength: 100
    DurationMinutes:
      type: integer
      minimum: 1
      maximum: 1440
    ErrorBody:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        seatsLeft:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        maxSeats:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
      required:
        - error
      additionalProperties: {}
...
    GuestTicketResponse:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/AuthTicket"
        messenger:
          $ref: "#/components/schemas/Messenger"
        messengerId:
          $ref: "#/components/schemas/MessengerID"
        displayName:
          type: string
      required:
        - ticket
        - messenger
        - messengerId
        - displayName
...
    InvalidBody:
      type: object
      properties:
        error:
          type: string
        details:
          $ref: "#/components/schemas/ValidationErrors"
      required:
        - error
        - details
...
    Location:
      type: string
      minLength: 1
      maxLength: 300
...
    MaxSeatsPerBooking:
      type: integer
      minimum: 1
      maximum: 1000
    Messenger:
      type: string
      enum:
        - telegram
    MessengerID:
      type: string
      minLength: 1
      maxLength: 100
...
    OptionLabel:
      type: string
      minLength: 1
      maxLength: 100
    OptionsList:
      minItems: 1
      maxItems: 50
      type: array
      items:
        $ref: "#/components/schemas/OptionLabel"
    OptionsSelectMode:
      type: string
      enum:
        - single
        - multi
...
    PriceText:
      type: string
      minLength: 1
      maxLength: 50
...
    ServiceDescription:
      type: string
      maxLength: 2000
    ServiceEnvelope:
      type: object
      properties:
        service:
          $ref: "#/components/schemas/ServiceRecord"
      required:
        - service
    ServiceID:
      type: string
      pattern: ^[A-Za-z0-9_-]{6,32}$
...
    ServiceRecord:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ServiceID"
        organizerId:
          $ref: "#/components/schemas/UUID"
        title:
          $ref: "#/components/schemas/DisplayName"
        description:
          anyOf:
            - type: string
            - type: "null"
        photoUrl:
          anyOf:
            - type: string
            - type: "null"
        location:
          anyOf:
            - type: string
            - type: "null"
        contact:
          anyOf:
            - type: string
            - type: "null"
        defaultPrice:
          $ref: "#/components/schemas/PriceText"
        defaultCapacity:
          $ref: "#/components/schemas/Capacity"
        defaultDurationMinutes:
          $ref: "#/components/schemas/DurationMinutes"
        maxSeatsPerBooking:
          $ref: "#/components/schemas/MaxSeatsPerBooking"
        options:
          anyOf:
            - type: array
              items:
                type: string
            - type: "null"
        optionsSelectMode:
          anyOf:
            - $ref: "#/components/schemas/OptionsSelectMode"
            - type: "null"
        createdAt:
          type: string
      required:
        - id
        - organizerId
        - title
        - description
        - photoUrl
        - location
        - contact
        - defaultPrice
        - defaultCapacity
        - defaultDurationMinutes
        - maxSeatsPerBooking
        - options
        - optionsSelectMode
        - createdAt
    ServicesEnvelope:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRecord"
      required:
        - services
...
    TelegramAuthDate:
      type: integer
      minimum: 1
      maximum: 4102444800
    TelegramHash:
      type: string
      minLength: 64
      maxLength: 64
    TelegramName:
      type: string
      minLength: 1
      maxLength: 256
    TelegramOptionalName:
      type: string
      maxLength: 256
    TelegramUserID:
      type: integer
      minimum: 1
      maximum: 9007199254740991
    TelegramWidgetPayload:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/TelegramUserID"
        first_name:
          $ref: "#/components/schemas/TelegramName"
        last_name:
          $ref: "#/components/schemas/TelegramOptionalName"
        username:
          $ref: "#/components/schemas/TelegramOptionalName"
        photo_url:
          $ref: "#/components/schemas/URL"
        auth_date:
          $ref: "#/components/schemas/TelegramAuthDate"
        hash:
          $ref: "#/components/schemas/TelegramHash"
      required:
        - id
        - first_name
        - auth_date
        - hash
...
    URL:
      type: string
      format: uri
    UUID:
      type: string
      format: uuid
      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
...
    ValidationErrors:
      type: object
      properties:
        formErrors:
          type: array
          items:
            type: string
        fieldErrors:
          type: object
          propertyNames:
            type: string
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - formErrors
        - fieldErrors